Adding validation to install.sh

+++ install.sh
@@ -24,8 +24,11 @@ echo "...done"
 
 # move any existing dotfiles in homedir to dotfiles_old directory, then create symlinks 
 for file in $files; do
+  if [ -f ~/.$file ] || [ -d ~/.$file ]
+  then
     echo "Moving any existing dotfiles from ~ to $olddir"
     mv ~/.$file ~/dotfiles_old/
-    echo "Creating symlink to $file in home directory."
-    ln -s $dir/$file ~/.$file
+  fi
+  echo "Creating symlink to $file in home directory."
+  ln -s $dir/$file ~/.$file
 done
